backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
openai
python-dotenv
python-multipart
